app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
kafka-python==2.0.2
sendgrid==6.9.2
psycopg2-binary==2.9.1
confluent-kafka>=1.0.0
